% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_lat_long.R
\name{get_lat_long}
\alias{get_lat_long}
\title{get_lat_long
Return lattitude and longitude from an address and zipcode
Queries the Utah UGRC API to obtain lat/long}
\usage{
get_lat_long(address, zipcode, apikey = NULL)
}
\arguments{
\item{address}{The street address to geocode}

\item{zipcode}{The zipcode associated with the `address`}

\item{apikey}{UGRC authenticaion key. System key used as default.}
}
\value{
Named list of latitude and longitude in NAD83
}
\description{
Handles authentication and connecting to UGRC API
to return lat/long from an address
}
\examples{
get_lat_long(address = '3286 E Vera Cir', zipcode = 84121)


}
