% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_lat_long.R
\name{get_lat_long}
\alias{get_lat_long}
\title{get_lat_long
Create an API connection tot he UGRC API}
\usage{
get_lat_long(address, zipcode, apikey = NULL)
}
\arguments{
\item{address}{The street address to geocode}

\item{zipcode}{The zipcode associated with the `address`}

\item{apikey}{UGRC authenticaion key. System key used as default.}
}
\value{
Authenticated connection
}
\description{
Handles authentication and connecting to UGRC API
}
\examples{
get_lat_long(address = '3286 E Vera Cir', zipcode = 84121)


}
